refactor: move prometheus metric out of server.cpp
